+++ src/styles/minipage.css
